pyproject.toml
[project]
name = "homeagent"
version = "0.1.0"
description = "Smart Home AI backbone"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "openai-whisper>=20250625",
    "pvporcupine>=4.0.1",
    "pyaudio>=0.2.14",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
